Seperate out config from conda-store functions
